Split out image code and add unit test
